split to assert.cpp comments
