+++ src/utility/fs.h
@@ -5,6 +5,7 @@ namespace mob
 
 class context;
 
+
 struct handle_closer
 {
 	using pointer = HANDLE;
@@ -31,6 +32,8 @@ struct file_closer
 using file_ptr = std::unique_ptr<FILE, file_closer>;
 
 
+// deletes the given file in the destructor unless cancel() is called
+//
 class file_deleter
 {
 public:
@@ -49,6 +52,8 @@ private:
 };
 
 
+// deletes the given directory in the destructor unless cancel is called
+//
 class directory_deleter
 {
 public:
@@ -67,15 +72,47 @@ private:
 };
 
 
+// creates a temporary file in the given directory that is used to detect
+// crashes or interruptions; some prefix is added to the filename to try and
+// make it unlikely to clash
+//
+// for example:
+//
+//   interruption_file ifile("some/dir", "some action");
+//
+//   if (ifile.exists())
+//   {
+//      // action was previously interrupted, do something about it
+//   }
+//
+//   // create interruption file
+//   ifile.create();
+//
+//   // do stuff that might fail and throw or return early
+//
+//   // success, remove
+//   ifile.remove();
+//
 class interruption_file
 {
 public:
 	interruption_file(const context& cx, fs::path dir, std::string name);
 
+	// path to the interruption file
+	//
 	fs::path file() const;
+
+	// whether the file exists
+	//
 	bool exists() const;
 
+
+	// creates the interruption file
+	//
 	void create();
+
+	// removes the interruption file
+	//
 	void remove();
 
 private:
@@ -85,12 +122,36 @@ private:
 };
 
 
+// creates a file in the given directory that is used to bypass an operation
+// in the future; some prefix is added to the filename to try and make it
+// unlikely to clash
+//
+// for example:
+//
+//   bypass_file built(cx, "some/dir/", "built");
+//
+//   if (built.exists())
+//   {
+//      // already built, bypass
+//      return;
+//   }
+//
+//   // do the build process
+//
+//   // bypass next time
+//   built.create();
+//
 class bypass_file
 {
 public:
 	bypass_file(const context& cx, fs::path dir, std::string name);
 
+	// whether the bypass file exists
+	//
 	bool exists() const;
+
+	// creates the bypass file
+	//
 	void create();
 
 private:
